@@ -1,47 +1,28 @@
 {
-  "_args": [
-    [
-      {
-        "raw": "iconv-lite@0.4.5",
-        "scope": null,
-        "escapedName": "iconv-lite",
-        "name": "iconv-lite",
-        "rawSpec": "0.4.5",
-        "spec": "0.4.5",
-        "type": "version"
-      },
-      "/Users/macos/myCode/20170815up-Archive/node_modules/body-parser"
-    ]
-  ],
-  "_from": "iconv-lite@0.4.5",
-  "_id": "iconv-lite@0.4.5",
-  "_inCache": true,
+  "_from": "iconv-lite@0.4.19",
+  "_id": "iconv-lite@0.4.19",
+  "_inBundle": false,
+  "_integrity": "sha512-oTZqweIP51xaGPI4uPa56/Pri/480R+mo7SeU+YETByQNhDG55ycFyNLIgta9vXhILrxXDmF7ZGhqZIcuN0gJQ==",
   "_location": "/iconv-lite",
-  "_nodeVersion": "0.10.33",
-  "_npmUser": {
-    "name": "ashtuchkin",
-    "email": "ashtuchkin@gmail.com"
-  },
-  "_npmVersion": "2.1.6",
   "_phantomChildren": {},
   "_requested": {
-    "raw": "iconv-lite@0.4.5",
-    "scope": null,
-    "escapedName": "iconv-lite",
+    "type": "version",
+    "registry": true,
+    "raw": "iconv-lite@0.4.19",
     "name": "iconv-lite",
-    "rawSpec": "0.4.5",
-    "spec": "0.4.5",
-    "type": "version"
+    "escapedName": "iconv-lite",
+    "rawSpec": "0.4.19",
+    "saveSpec": null,
+    "fetchSpec": "0.4.19"
   },
   "_requiredBy": [
     "/body-parser",
     "/raw-body"
   ],
-  "_resolved": "https://registry.npmjs.org/iconv-lite/-/iconv-lite-0.4.5.tgz",
-  "_shasum": "9c574b70c30d615859f2064d2be4335ad6b1a8d6",
-  "_shrinkwrap": null,
-  "_spec": "iconv-lite@0.4.5",
-  "_where": "/Users/macos/myCode/20170815up-Archive/node_modules/body-parser",
+  "_resolved": "https://registry.npmjs.org/iconv-lite/-/iconv-lite-0.4.19.tgz",
+  "_shasum": "f7468f60135f5e5dad3399c0a81be9a1603a082b",
+  "_spec": "iconv-lite@0.4.19",
+  "_where": "/Users/lekhanhthong/WebstormProjects/joboana/node_modules/body-parser",
   "author": {
     "name": "Alexander Shtuchkin",
     "email": "ashtuchkin@gmail.com"
@@ -53,6 +34,7 @@
   "bugs": {
     "url": "https://github.com/ashtuchkin/iconv-lite/issues"
   },
+  "bundleDependencies": false,
   "contributors": [
     {
       "name": "Jinwu Zhan",
@@ -103,26 +85,21 @@
       "url": "https://github.com/nleush"
     }
   ],
-  "dependencies": {},
+  "deprecated": false,
   "description": "Convert character encodings in pure javascript.",
   "devDependencies": {
     "async": "*",
     "errto": "*",
-    "iconv": "~2.1.4",
+    "iconv": "*",
     "istanbul": "*",
     "mocha": "*",
-    "request": "2.47",
+    "request": "*",
+    "semver": "*",
     "unorm": "*"
   },
-  "directories": {},
-  "dist": {
-    "shasum": "9c574b70c30d615859f2064d2be4335ad6b1a8d6",
-    "tarball": "https://registry.npmjs.org/iconv-lite/-/iconv-lite-0.4.5.tgz"
-  },
   "engines": {
-    "node": ">=0.8.0"
+    "node": ">=0.10.0"
   },
-  "gitHead": "0654719791aa2c159bb3820f095b5da8702d091b",
   "homepage": "https://github.com/ashtuchkin/iconv-lite",
   "keywords": [
     "iconv",
@@ -132,15 +109,7 @@
   ],
   "license": "MIT",
   "main": "./lib/index.js",
-  "maintainers": [
-    {
-      "name": "ashtuchkin",
-      "email": "ashtuchkin@gmail.com"
-    }
-  ],
   "name": "iconv-lite",
-  "optionalDependencies": {},
-  "readme": "ERROR: No README data found!",
   "repository": {
     "type": "git",
     "url": "git://github.com/ashtuchkin/iconv-lite.git"
@@ -150,5 +119,6 @@
     "coverage-open": "open coverage/lcov-report/index.html",
     "test": "mocha --reporter spec --grep ."
   },
-  "version": "0.4.5"
+  "typings": "./lib/index.d.ts",
+  "version": "0.4.19"
 }
